Config 2.0: Lazy.nvim done right
Nvim config 2.0

This update took a lot of effort to make as it was completely built from the ground up, but oh boy was it worth it!

- LSP

The biggest change is that LSP for python is setup the right way now! This was one of the main goals of this update as the previous iteration was not clear for me and I wasn't happy with it.

ruff-lsp: Ruff is now in charge of providing the diagnostics for my python code and it is BLAZINGLY FAST. It also does formatting when I save the file.

pylsp: I still use pylsp to do Hover actions, signature help and navigating through my sources.

rezolve: I wrote this custom script to be able to use these tool which would only be available inside of a rez context in my setup, that way I don't have to jump into a rez environment to use it.

- Completion

Nvim-cmp is now setup in a super nice way, no more fighting with tab and at the end of a paragraph, things just work.

Integration with Luasnip! I didn't get the power of snippets until I started to use them and I will never go back.

- QOL

Added treesitter text objects and incremental selection.
Added bufferline to quickly view and manage currently open buffers.
Added comment.nvim to quickly add comments for any language.
Added which-key to never forget anything :).
Added ultimate-autopairs.nvim and wow! it's amazing how much it improves my workflow.
And many, many more new features. Some things are still missing from my previous config but these are all the basics I needed to supercharge my nvim experience :)